Further refactoring to make setup steps independant.
